skills/multi-writing-skills/pyproject.toml
[project]
name = "multi-writing-skills"
version = "0.1.0"
description = "多平台 Markdown 发布工具，支持微信公众号、知乎、今日头条"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "yuesf", email = "yuesf@example.com" }
]
keywords = ["writing", "markdown", "publisher", "wechat", "zhihu", "toutiao"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "Pillow>=10.0",
    "markdown-it-py>=3.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "rich>=13.0",
    "linkify-it-py>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "black>=24.0",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
multi-writing-skills = "multi_writing_skills.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/multi_writing_skills"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
